fixing bugs

+++ frontend/src/app/board/list-board/list-board.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { BoardService } from '../../services/board.service';
 import { OwlOptions } from 'ngx-owl-carousel-o';
 import { ActivatedRoute, Params, Router } from '@angular/router';
+import Swal from 'sweetalert2'
 import {
   MatSnackBar,
   MatSnackBarHorizontalPosition,
@@ -43,20 +44,49 @@ export class ListBoardComponent implements OnInit {
   }
 
   deleteBoard(board: any) {
-    this._boardService.deleteBoard(board).subscribe(
-      (res) => {
-        let index = this.boardData.indexOf(board);
-        if (index > -1) {
-          this.boardData.splice(index, 1);
-          this.message = res.message;
-          this.openSnackBarSuccesfull();
-        }
-      },
-      (err) => {
-        this.message = err.error;
-        this.openSnackBarError();
+    
+    Swal.fire({
+      title: 'Are you sure you want to unsubscribe the task?',
+      text: 'The task will be assigned to the administrator and then assigned to a board member.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, unsubscribe!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._boardService.deleteBoard(board).subscribe(
+          (res) => {
+            let index = this.boardData.indexOf(board);
+            if (index > -1) {
+              this.boardData.splice(index, 1);
+              this.message = res.message;
+              this.openSnackBarSuccesfull();
+              Swal.fire(
+                this.message,
+                'Board Deleted.',
+                'success'
+              );
+            }
+          },
+          (err)=>{
+            this.message = err.error;
+            Swal.fire(
+              this.message,
+              'Cant Delete the board',
+              'error'
+            );
+            console.log(err.error)
+          }
+        )
+        
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'You cancel the process have a nice day',
+          'error'
+        )
       }
-    );
+    })
   }
   updateBoard(board: any) {
     this._boardService.deleteBoard(board).subscribe(
@@ -65,6 +95,11 @@ export class ListBoardComponent implements OnInit {
         if (index > -1) {
           this.boardData.splice(index, 1);
           this.message = res.message;
+          Swal.fire(
+            this.message,
+            'Board Updated.',
+            'success'
+          );
           this.openSnackBarSuccesfull();
         }
       },
